% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{CarrierEnableDisable}
\alias{CarrierEnableDisable}
\alias{EnableCarrier}
\alias{DisableCarrier}
\title{Enable or disable carrier output}
\usage{
EnableCarrier(device, channel)

DisableCarrier(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Enable or disable the carrier node for an Analog Out channel.

\code{EnableCarrier()} enables the carrier node.

\code{DisableCarrier()} disables the carrier node.
}
\details{
These functions change the carrier node configuration. To apply the
  changes, call \code{ApplyAnalogOutSettings()} or \code{StartAnalogOut()}.
}
\seealso{
See the vignette \emph{Basic Analog Out Functionality} for a complete
  Analog Out workflow.
}
