% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_management.R
\name{CloseAllDevices}
\alias{CloseAllDevices}
\title{Close all WaveForms devices opened by the current process}
\usage{
CloseAllDevices()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Close all WaveForms device connections opened by the current R process.
  This operation is process-wide and is not limited to connections opened
  through \pkg{dwf4r}. It may therefore also close connections created by
  other packages or by direct use of the WaveForms SDK in the same R process.
  Connections owned by other processes are not affected.

  After this function returns, all existing \code{"dwf4r_device"} objects
  and their copies are stale and cannot be used in subsequent \pkg{dwf4r}
  calls. Any WaveForms handles created by other code in the current R
  process are also invalidated.
}
\examples{
  \dontrun{
    devices <- ListDevices()
    if (nrow(devices) >= 1L) {
      device_sn <- devices$serial_number[[1L]]
      device <- OpenDevice(device_sn)
      CloseAllDevices()
    }
  }

}
