% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_management.R
\name{CloseDevice}
\alias{CloseDevice}
\title{Close a device}
\usage{
CloseDevice(device)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Close an active WaveForms device connection and mark the corresponding
  session as closed. After this function returns, the \code{device} object
  and any copies of it are stale and cannot be used in subsequent \pkg{dwf4r}
  calls.
}
\examples{
  \dontrun{
    devices <- ListDevices()
    if (nrow(devices) >= 1L) {
      device_sn <- devices$serial_number[[1L]]
      device <- OpenDevice(device_sn)
      CloseDevice(device)
    }
  }

}
