% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.dwf_constants}
\alias{.dwf_constants}
\title{Internal WaveForms SDK constants}
\format{
A named list.
}
\usage{
.dwf_constants
}
\description{
Integer mappings between user-facing names and WaveForms SDK constants.
}
\keyword{internal}
