% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{GetCarrierFunction}
\alias{GetCarrierFunction}
\title{Get the carrier waveform function}
\usage{
GetCarrierFunction(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
A string identifying the configured waveform function. Possible values can
  be queried with \code{GetAnalogOutNodeFunctionTypes()}.
}
\description{
Retrieve the waveform function currently configured for the carrier node
  of an Analog Out channel.
}
\seealso{
See the vignette \emph{Basic Analog Out Functionality} for a complete
  Analog Out workflow.
}
