% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{GetCarrierSettings}
\alias{GetCarrierSettings}
\title{Get carrier waveform settings}
\usage{
GetCarrierSettings(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
A named list with the following elements:
  \describe{
    \item{\code{enabled}}{
      Logical value indicating whether the carrier node is enabled.
    }
    \item{\code{func}}{
      String identifying the configured waveform function.
    }
    \item{\code{frequency}}{
      Numeric carrier frequency in Hz.
    }
    \item{\code{amplitude}}{
      Numeric carrier peak amplitude in volts relative to the configured
      offset.
    }
    \item{\code{offset}}{
      Numeric carrier offset in volts.
    }
    \item{\code{symmetry}}{
      Numeric carrier symmetry in percent.
    }
    \item{\code{phase}}{
      Numeric carrier phase in degrees.
    }
  }
}
\description{
Retrieve the current configuration of the carrier node for an Analog Out
  channel.
}
\seealso{
See the vignette \emph{Basic Analog Out Functionality} for a complete
  Analog Out workflow.
}
