% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{IsCarrierEnabled}
\alias{IsCarrierEnabled}
\title{Check whether carrier output is enabled}
\usage{
IsCarrierEnabled(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
A logical scalar, \code{TRUE} if the carrier node is enabled and
  \code{FALSE} if it is disabled.
}
\description{
Retrieve the configured enable state of the carrier node for an Analog Out
  channel.
}
\seealso{
See the vignette \emph{Basic Analog Out Functionality} for a complete
  Analog Out workflow.
}
