% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_management.R
\name{OpenDevice}
\alias{OpenDevice}
\title{Open a device}
\usage{
OpenDevice(serial_number = NULL)
}
\arguments{
\item{serial_number}{A string containing the serial number of the device to open. If
\code{NULL}, the device is selected automatically when exactly one
compatible device is detected.}
}
\value{
An opaque object of class \code{"dwf4r_device"} representing the opened
  connection. Pass this object unchanged to other \pkg{dwf4r} functions.
}
\description{
Open a WaveForms device, register the connection in the current R session,
  and return an object that can be passed to other \pkg{dwf4r} functions.
}
\details{
The device is opened using its default configuration. Automatic
  configuration in the WaveForms SDK is disabled to reduce the communication
  overhead of individual setting calls. Instrument settings are therefore
  applied explicitly by the corresponding \pkg{dwf4r} configuration
  functions. This behavior is managed internally and is not currently
  user-configurable.
}
\examples{
  \dontrun{
    devices <- ListDevices()

    if (nrow(devices) >= 1L) {
      if (nrow(devices) == 1L) {
        device <- OpenDevice()
      } else {
        device_sn <- devices$serial_number[[1L]]
        device <- OpenDevice(device_sn)
      }

      print(device)

      CloseDevice(device)
    }
  }

}
