% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_info.R
\name{GetAnalogOutInfo}
\alias{GetAnalogOutInfo}
\alias{GetAnalogOutChannelCount}
\alias{GetAnalogOutChannelNodes}
\alias{GetAnalogOutRunRange}
\alias{GetAnalogOutWaitRange}
\alias{GetAnalogOutRepeatRange}
\alias{GetAnalogOutIdleModes}
\alias{GetAnalogOutNodeFunctionTypes}
\alias{GetAnalogOutNodeFrequencyRange}
\alias{GetAnalogOutNodeAmplitudeRange}
\alias{GetAnalogOutNodeOffsetRange}
\alias{GetAnalogOutNodeSymmetryRange}
\alias{GetAnalogOutNodePhaseRange}
\alias{GetAnalogOutNodeSampleCountRange}
\title{Get Analog Out characteristics}
\usage{
GetAnalogOutChannelCount(device)

GetAnalogOutChannelNodes(device, channel)

GetAnalogOutRunRange(device, channel)

GetAnalogOutWaitRange(device, channel)

GetAnalogOutRepeatRange(device, channel)

GetAnalogOutIdleModes(device, channel)

GetAnalogOutNodeFunctionTypes(device, channel, node)

GetAnalogOutNodeFrequencyRange(device, channel, node)

GetAnalogOutNodeAmplitudeRange(device, channel, node)

GetAnalogOutNodeOffsetRange(device, channel, node)

GetAnalogOutNodeSymmetryRange(device, channel, node)

GetAnalogOutNodePhaseRange(device, channel, node)

GetAnalogOutNodeSampleCountRange(device, channel, node)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}

\item{node}{A string specifying the Analog Out node. One of \code{"carrier"},
\code{"fm"}, or \code{"am"}.}
}
\value{
An integer, numeric, or character vector describing the requested Analog
  Out capability. Range functions return two elements containing the minimum
  and maximum supported values.
}
\description{
Retrieve Analog Out characteristics of a device.

\code{GetAnalogOutChannelCount()} returns the number of Analog Out
  channels.

\code{GetAnalogOutChannelNodes()} returns the nodes supported by an Analog
  Out channel.

\code{GetAnalogOutRunRange()} returns the supported run-time range, in
  seconds, for an Analog Out channel.

\code{GetAnalogOutWaitRange()} returns the supported wait-time range, in
  seconds, for an Analog Out channel.

\code{GetAnalogOutRepeatRange()} returns the supported repeat-count range
  for an Analog Out channel.

\code{GetAnalogOutIdleModes()} returns the idle output modes supported by
  an Analog Out channel.

\code{GetAnalogOutNodeFunctionTypes()} returns the waveform functions
  supported by an Analog Out node.

\code{GetAnalogOutNodeFrequencyRange()} returns the supported frequency
  range for an Analog Out node.

\code{GetAnalogOutNodeAmplitudeRange()} returns the supported amplitude
  range for an Analog Out node.

\code{GetAnalogOutNodeOffsetRange()} returns the supported offset range for
  an Analog Out node.

\code{GetAnalogOutNodeSymmetryRange()} returns the supported symmetry range
  for an Analog Out node.

\code{GetAnalogOutNodePhaseRange()} returns the supported phase range, in
  degrees, for an Analog Out node.

\code{GetAnalogOutNodeSampleCountRange()} returns the supported range of
  sample counts for custom waveform data.
}
\seealso{
See the vignette \emph{Basic Analog Out Functionality} for a complete
  Analog Out workflow.
}
