% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{GetAnalogOutRepeatTrigger}
\alias{GetAnalogOutRepeatTrigger}
\title{Get Analog Out repeat triggering}
\usage{
GetAnalogOutRepeatTrigger(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
A logical scalar. \code{TRUE} indicates that a new trigger is included in
  each wait-run repeat cycle; \code{FALSE} indicates that repeated cycles
  proceed without waiting for another trigger.
}
\description{
Retrieve whether a new trigger is included in each repeat cycle of a
  selected Analog Out channel.
}
