% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{GetAnalogOutStatus}
\alias{GetAnalogOutStatus}
\title{Get the state of an Analog Out channel}
\usage{
GetAnalogOutStatus(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
A string containing \code{"ready"}, \code{"armed"}, \code{"wait"},
  \code{"running"}, \code{"done"}, \code{"config"}, \code{"prefill"}, or
  \code{"not_done"}. The value \code{"unknown"} is returned for an
  unrecognized SDK state.
}
\description{
Retrieve the current state of a selected Analog Out channel.
}
\seealso{
See the vignette \emph{Basic Analog Out Functionality} for a complete
  Analog Out workflow.
}
