% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{GetAnalogOutTriggerSlope}
\alias{GetAnalogOutTriggerSlope}
\title{Get the Analog Out trigger slope}
\usage{
GetAnalogOutTriggerSlope(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
A string identifying the configured trigger slope.
}
\description{
Retrieve the trigger slope currently configured for a selected Analog Out
  channel.
}
