% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{GetAnalogOutTriggerSource}
\alias{GetAnalogOutTriggerSource}
\title{Get the Analog Out trigger source}
\usage{
GetAnalogOutTriggerSource(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
A string identifying the configured trigger source.
}
\description{
Retrieve the trigger source currently configured for a selected Analog Out
  channel.
}
