% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_out_configure.R
\name{ResetAnalogOut}
\alias{ResetAnalogOut}
\title{Reset Analog Out channel}
\usage{
ResetAnalogOut(device, channel)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{channel}{An integer scalar representing the zero-based Analog Out channel index.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Reset the Analog Out settings of a selected channel to their SDK defaults.
  Because automatic configuration is disabled by \pkg{dwf4r}, call
  \code{ApplyAnalogOutSettings()} or \code{StartAnalogOut()} to apply the
  reset configuration to the device.
}
\seealso{
See the vignette \emph{Basic Analog Out Functionality} for a complete
  Analog Out workflow.
}
