% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital_out_configure.R
\name{ResetDigitalOut}
\alias{ResetDigitalOut}
\title{Reset Digital Out instrument}
\usage{
ResetDigitalOut(device, .validate_device = TRUE)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}

\item{.validate_device}{A logical scalar indicating whether \code{AssertDevice()} should be called.
This argument is intended for advanced use to avoid repeated validation in
controlled sequences of operations. Set to \code{FALSE} only when the
device has already been validated and is known to remain open. Other
argument validation and SDK error handling remain active.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Reset all Digital Out instrument parameters to their SDK defaults. Because
  automatic configuration is disabled by \pkg{dwf4r}, call
  \code{StartDigitalOut()} or \code{StopDigitalOut()} after resetting,
  depending on the desired instrument state.
}
\seealso{
See the vignette \emph{Basic Digital Out Functionality} for a complete
  Digital Out workflow.
}
