% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_trigger.R
\name{TriggerDevice}
\alias{TriggerDevice}
\title{Trigger a WaveForms device}
\usage{
TriggerDevice(device)
}
\arguments{
\item{device}{A \code{"dwf4r_device"} object returned by \code{OpenDevice()} that refers
to a device connection that is still open.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Generate a PC trigger pulse for a WaveForms device.

  Instruments configured to use the \code{"pc"} trigger source can respond
  to this pulse.
}
